requirements.txt
pandas>=3.0
pyarrow>=14.0
matplotlib>=3.9
seaborn>=0.13
jupyter>=1.0
python>=3.12
numpy>=2.0
ipykernel>=6.0